Improved oauth workflow for new users

+++ private/system/lib-security.php
@@ -494,7 +494,7 @@ function SEC_getUserPermissions($grp_id='',$uid='')
         $groups = SEC_getUserGroups ($uid);
     }
 
-    if ( count($groups) > 0 ) {
+    if ( is_array($groups) && count($groups) > 0 ) {
         $glist = join(',', $groups);
         $result = DB_query("SELECT DISTINCT ft_name FROM {$_TABLES["access"]},{$_TABLES["features"]} "
                          . "WHERE ft_id = acc_ft_id AND acc_grp_id IN ($glist)");
@@ -1120,7 +1120,7 @@ function SEC_createToken($ttl = TOKEN_TTL)
 
     static $_tokenKey;
 
-    if ( $ttl == -1 || COM_isAnonUser() ) {
+    if ( $ttl == -1 ) {
         $tokenKey = '';
         return;
     }
@@ -1148,9 +1148,10 @@ function SEC_createToken($ttl = TOKEN_TTL)
     DB_query($sql);
 
     /* Destroy tokens for this user/url combination */
-    $sql = "DELETE FROM {$_TABLES['tokens']} WHERE owner_id={$uid} AND urlfor='".DB_escapeString($pageURL)."'";
-    DB_query($sql);
-
+    if ( !COM_isAnonUser()) {
+        $sql = "DELETE FROM {$_TABLES['tokens']} WHERE owner_id={$uid} AND urlfor='".DB_escapeString($pageURL)."'";
+        DB_query($sql);
+    }
     /* Create a token for this user/url combination */
     /* NOTE: TTL mapping for PageURL not yet implemented */
     $sql = "INSERT INTO {$_TABLES['tokens']} (token, created, owner_id, urlfor, ttl) "
@@ -1182,8 +1183,6 @@ function SEC_checkToken()
         SEC_createToken(-1);
         return true;
     }
-
-    if ( COM_isAnonUser() ) return false;
 
     if ( !SEC_isLocalUser($_USER['uid']) ) {
         return false;
@@ -1251,8 +1250,6 @@ function _sec_checkToken($ajax=0)
 
     $token = ''; // Default to no token.
     $return = false; // Default to fail.
-
-    if ( COM_isAnonUser() ) return true;
 
     if ( isset($_SYSTEM['token_ip']) && $_SYSTEM['token_ip'] == true ) {
         $referCheck  = $_SERVER['REMOTE_ADDR'];
